Move strings out of riscv-meta into riscv-strings
- Rename string tables to be consistent with their associated
  enum with the addition of the '_sym' suffix:

  - Rename riscv_i_registers to riscv_ireg_name_sym
  - Rename riscv_f_registers to riscv_freg_name_sym
  - Rename riscv_insn_name to riscv_insn_name_sym

- Print instruction mask and match tables using PRIx64
  to make provision for 6 byte and 8 byte instructions.

- Change riscv_opcode match and mask elements from
  size_t to uint64_t as we specifically require 64-bits.

- Added #ifdef __cplusplus extern "C" to riscv-meta and
  riscv-strings headers (the metadata uses C linkage).

- Change formattting of compression and argument metadata.

+++ src/riscv-model.h
@@ -221,9 +221,9 @@ struct riscv_opcode
 	riscv_compressed_list compressions;
 
 	size_t num;
-	size_t mask;
-	size_t match;
-	size_t done;
+	uint64_t mask;
+	uint64_t match;
+	uint64_t done;
 
 	riscv_opcode(std::string key, std::string name)
 		: key(key), name(name), num(0), mask(0), match(0), done(0) {}
